+++ WeightliftingTeam1/WeightliftingTeam1/Components/GridForUsers.razor
@@ -1,0 +1,15 @@
+﻿@typeparam TItem
+
+<TelerikGrid Data=GridData
+             @bind-Page=CurrentPage PageSize=PageSize
+             Height="100%"
+             Pageable="true" Sortable="true" Groupable="false"
+             Resizable="true" Reorderable="true" AutoGenerateColumns="true"
+             Width="100%">
+    <GridToolBar>
+        <GridSearchBox />
+    </GridToolBar>
+    <GridColumns>
+        <GridAutoGeneratedColumns ColumnWidth="110px" />
+    </GridColumns>
+</TelerikGrid>
