--- WeightliftingTeam1/WeightliftingTeam1/Components/MyTelerikGrid.razor
@@ -1,75 +1,0 @@
-﻿@using WeightliftingTeam1.Data
-@using WeightliftingTeam1.ModelsForOutput
-@using WeightliftingTeam1.Panels
-
-
-@if (DataForGrids == null)
-{
-    <p>Loading @PanelType......</p>
-}
-else
-{
-    @switch (PanelType)
-    {
-        case PanelType.Attempts:
-            <TelerikGrid Data=@DataForGrids.Attempts
-                         @bind-Page=CurrentPage PageSize=PageSize
-                         Height="100%"
-                         Pageable="true" Sortable="true" Groupable="false"
-                         Resizable="true" Reorderable="true" AutoGenerateColumns="true"
-                         Width="100%">
-                <GridToolBar>
-                    @*<AuthorizeView>
-                        <Authorized>
-                            <GridCommandButton Command="Add" Icon="add">Add Item</GridCommandButton>
-                        </Authorized>
-                    </AuthorizeView>*@
-                    <GridSearchBox />
-                </GridToolBar>
-                <GridColumns>
-                    <GridAutoGeneratedColumns ColumnWidth="110px" />
-                </GridColumns>
-            </TelerikGrid>
-            break;
-
-        case PanelType.Athletes:
-
-            <TelerikGrid Data="@DataForGrids.Athletes"
-                         @bind-Page=CurrentPage PageSize=PageSize
-                         Height="100%"
-                         Pageable="true" Sortable="true" Groupable="false"
-                         Resizable="false" Reorderable="true" AutoGenerateColumns="true"
-                         Width="100%">
-                <GridToolBar>
-                    <GridSearchBox />
-                </GridToolBar>
-                <GridColumns>
-                    <GridAutoGeneratedColumns ColumnWidth="110px" />
-                </GridColumns>
-            </TelerikGrid>
-            break;
-        case PanelType.Records:
-            <TelerikGrid Data=@DataForGrids.Records
-                         @bind-Page=CurrentPage PageSize=PageSize
-                         Height="100%"
-                         Pageable="true" Sortable="true" Groupable="false"
-                         Resizable="false" Reorderable="true" AutoGenerateColumns="true"
-                         Width="100%">
-                <GridToolBar>
-                    <GridSearchBox />
-                </GridToolBar>
-                <GridColumns>
-                    <GridAutoGeneratedColumns ColumnWidth="110px" />
-                </GridColumns>
-            </TelerikGrid>
-            break;
-    }
-}
-
-<style>
-    @@media (max-width: 1050px) {
-        .k-pager-numbers-wrap {
-            display: none;
-        }
-    }
-</style>
