+++ WeightliftingTeam1/WeightliftingTeam1/Components/MyTelerikGrids.razor
@@ -1,0 +1,34 @@
+﻿@using WeightliftingTeam1.Data
+@using WeightliftingTeam1.ModelsForOutput
+@using WeightliftingTeam1.Panels
+
+
+@if (DataForGrids == null)
+{
+    <p>Loading @PanelType......</p>
+}
+else
+{
+    @switch (PanelType)
+    {
+        case PanelType.Attempts:
+            <GridForUsers GridData="DataForGrids.Attempts"></GridForUsers>
+            break;
+
+        case PanelType.Athletes:
+            <GridForUsers GridData="DataForGrids.Attempts"></GridForUsers>
+            break;
+
+        case PanelType.Records:
+            <GridForUsers GridData="DataForGrids.Attempts"></GridForUsers>
+            break;
+    }
+}
+
+<style>
+    @@media (max-width: 1050px) {
+        .k-pager-numbers-wrap {
+            display: none;
+        }
+    }
+</style>
