+++ WeightliftingTeam1/WeightliftingTeam1/Pages/Index.razor
@@ -60,7 +60,7 @@
                         </div>
                     </div>
                     <div class="col-md-8 col-lg-9 p-2">
-                        <MyTelerikGrid PanelType="CurrPanelType" DataForGrids="DataForGrids"></MyTelerikGrid>
+                        <MyTelerikGrids PanelType="CurrPanelType" DataForGrids="DataForGrids"></MyTelerikGrids>
                     </div>
                 </div>
             </div>
